ci: move the getters to the getter file
