+++ configs/src/deployments/index.ts
@@ -1,3 +1,4 @@
 export * from './types';
 export * from './getters';
 export * from './contracts';
+export * from './osx-ens-domains';
